refactor: move code from hedera-evm to hedera-evm-api
