+++ hedera-node/hedera-evm-api/src/main/java/com/hedera/services/evm/contracts/execution/HederaEvmTransactionProcessingResult.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hedera.services.evm.implementation.contracts.execution;
+package com.hedera.services.evm.contracts.execution;
 
 import java.util.Collections;
 import java.util.List;
