+++ hedera-node/hedera-evm-api/src/main/java/com/hedera/services/evm/contracts/execution/exceptions/InvalidTransactionException.java
@@ -13,15 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hedera.services.evm.implementation.exceptions;
+package com.hedera.services.evm.contracts.execution.exceptions;
 
 import com.hederahashgraph.api.proto.java.ResponseCodeEnum;
 import org.apache.tuweni.bytes.Bytes;
 
 /**
- * Captures a failure in transaction processing to be captured by the {@link
- * com.hedera.services.txns.TransitionRunner} and used to set the final resolved status of the
- * transaction.
  *
  * <p>Unless the contained {@link ResponseCodeEnum} is exactly {@code FAIL_INVALID}, this represents
  * some form of user error. The {@code FAIL_INVALID} code indicates an internal system error; and it
