+++ hedera-node/hedera-evm-api/src/main/java/module-info.java
@@ -6,6 +6,8 @@ module com.hedera.services.evm {
     requires com.hedera.hashgraph.protobuf.java.api;
     requires org.apache.commons.lang3;
     requires tuweni.bytes;
+    requires javax.inject.custom;
+    requires com.swirlds.common;
 
     exports com.hedera.services.evm;
     exports com.hedera.services.evm.contracts.execution;
