+++ hedera-node/hedera-evm/build.gradle.kts
@@ -21,13 +21,5 @@ group = "com.hedera.evm"
 description = "Hedera EVM - Implementation"
 
 dependencies {
-    api(libs.commons.lang3)
-    api(libs.protobuf.java)
-    api(libs.besu.evm)
-    api(libs.besu.datatypes)
-    api(libs.swirlds.common)
-    api(project(":hapi-utils"))
     api(project(":hedera-node:hedera-evm-api"))
-    api(project(":hedera-node:hedera-evm:javax-inject"))
-    implementation(libs.hapi)
 }
