+++ hedera-node/hedera-evm/src/main/java/module-info.java
@@ -1,11 +1,3 @@
 module com.hedera.services.evm.implementation {
-    requires org.hyperledger.besu.evm;
-    requires org.hyperledger.besu.datatypes;
-    requires com.google.protobuf;
     requires com.hedera.services.evm;
-    requires tuweni.bytes;
-    requires com.hedera.hashgraph.protobuf.java.api;
-    requires org.apache.commons.lang3;
-    requires com.swirlds.common;
-    requires javax.inject.custom;
 }
