+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/execution/CallLocalEvmTxProcessor.java
@@ -19,11 +19,10 @@ import static com.hedera.services.exceptions.ValidationUtils.validateTrue;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_CONTRACT_ID;
 
 import com.hedera.services.context.properties.GlobalDynamicProperties;
-import com.hedera.services.evm.implementation.contracts.execution.BlockMetaSource;
+import com.hedera.services.evm.contracts.execution.BlockMetaSource;
 import com.hedera.services.evm.store.contracts.HederaEvmMutableWorldState;
 import com.hedera.services.ledger.accounts.AliasManager;
 import com.hedera.services.store.contracts.CodeCache;
-import com.hedera.services.store.contracts.HederaMutableWorldState;
 import com.hedera.services.store.models.Account;
 import com.hederahashgraph.api.proto.java.HederaFunctionality;
 import java.util.Map;
