+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/execution/EvmTxProcessor.java
@@ -23,8 +23,8 @@ import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INSUFFICIENT_T
 
 import com.hedera.services.context.properties.GlobalDynamicProperties;
 import com.hedera.services.contracts.execution.traceability.HederaTracer;
-import com.hedera.services.evm.implementation.contracts.execution.BlockMetaSource;
-import com.hedera.services.evm.implementation.contracts.execution.HederaEvmTxProcessor;
+import com.hedera.services.evm.contracts.execution.BlockMetaSource;
+import com.hedera.services.evm.contracts.execution.HederaEvmTxProcessor;
 import com.hedera.services.exceptions.InvalidTransactionException;
 import com.hedera.services.exceptions.ResourceLimitException;
 import com.hedera.services.store.contracts.HederaMutableWorldState;
