+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/execution/StaticBlockMetaProvider.java
@@ -16,7 +16,7 @@
 package com.hedera.services.contracts.execution;
 
 import com.hedera.services.context.primitives.SignedStateViewFactory;
-import com.hedera.services.evm.implementation.contracts.execution.BlockMetaSource;
+import com.hedera.services.evm.contracts.execution.BlockMetaSource;
 import java.util.Optional;
 import javax.inject.Inject;
 import javax.inject.Singleton;
