+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/execution/StaticBlockMetaSource.java
@@ -15,7 +15,7 @@
  */
 package com.hedera.services.contracts.execution;
 
-import com.hedera.services.evm.implementation.contracts.execution.BlockMetaSource;
+import com.hedera.services.evm.contracts.execution.BlockMetaSource;
 import com.hedera.services.state.merkle.MerkleNetworkContext;
 import java.time.Instant;
 import org.hyperledger.besu.datatypes.Hash;
