+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/services/contracts/execution/TransactionProcessingResult.java
@@ -16,7 +16,7 @@
 package com.hedera.services.contracts.execution;
 
 import com.hedera.services.contracts.execution.traceability.SolidityAction;
-import com.hedera.services.evm.implementation.contracts.execution.HederaEvmTransactionProcessingResult;
+import com.hedera.services.evm.contracts.execution.HederaEvmTransactionProcessingResult;
 import com.hedera.services.state.submerkle.EvmFnResult;
 import com.hederahashgraph.api.proto.java.ContractFunctionResult;
 import com.hederahashgraph.api.proto.java.ContractID;
