+++ hedera-node/javax-inject/src/main/java/javax/inject/Scope.java
