+++ hedera-node/javax-inject/src/main/java/javax/inject/Singleton.java
